gsk: Adapt to moved function
