+++ gsk4/src/lib.rs
@@ -67,6 +67,9 @@ mod path;
 #[cfg(feature = "v4_14")]
 #[cfg_attr(docsrs, doc(cfg(feature = "v4_14")))]
 mod path_builder;
+#[cfg(feature = "v4_14")]
+#[cfg_attr(docsrs, doc(cfg(feature = "v4_14")))]
+mod path_point;
 mod radial_gradient_node;
 mod repeat_node;
 mod repeating_linear_gradient_node;
